Changes for working with entitlement change handler.
