+++ android/src/main/java/com/nami/reactlibrary/NamiUtil.kt
@@ -1,12 +1,10 @@
 package com.nami.reactlibrary
 
 import android.util.Log
-import com.facebook.react.bridge.Arguments
-import com.facebook.react.bridge.WritableArray
-import com.facebook.react.bridge.WritableMap
-import com.facebook.react.bridge.WritableNativeMap
+import com.facebook.react.bridge.*
 import com.namiml.paywall.NamiPaywall
 import com.namiml.billing.NamiPurchase
+import com.namiml.entitlement.NamiEntitlement
 import com.namiml.paywall.NamiSKU
 import java.text.DateFormat
 import java.text.SimpleDateFormat
@@ -203,6 +201,89 @@ fun purchaseToPurchaseDict(purchase: NamiPurchase): WritableMap {
 
     return purchaseMap
 }
+
+fun entitlementDictFromEntitlement(entitlement: NamiEntitlement): WritableMap? {
+    var resultMap: WritableMap = WritableNativeMap()
+    val referenceID: String = entitlement.referenceId ?: ""
+    resultMap.putString("referenceID", referenceID)
+
+    Log.i("NamiBridge", "Processing entitlement into Javascript Map with referenceID $referenceID")
+
+    if (referenceID.isEmpty()) {
+        // Without a reference ID, do not use this object
+        return null
+    }
+
+    val namiID: String = entitlement.namiId ?: ""
+    resultMap.putString("namiID", namiID)
+
+    val description: String = entitlement.desc ?: ""
+    resultMap.putString("desc", description)
+
+    val name: String = entitlement.name ?: ""
+    resultMap.putString("name", name)
+
+    val activePurchasesArray: WritableArray = WritableNativeArray()
+    val purchases = entitlement.activePurchases
+    purchases.let {
+        for (purchase in purchases) {
+            val purchaseMap = purchaseToPurchaseDict(purchase)
+            activePurchasesArray.pushMap(purchaseMap)
+        }
+    }
+    resultMap.putArray("activePurchases", activePurchasesArray)
+
+
+    val purchasedSKUsArray: WritableArray = WritableNativeArray()
+    val purchasedSKUs = entitlement.purchasedSKUs
+    for (sku in purchasedSKUs) {
+        val skuMap = skuToSkuDict(sku)
+        purchasedSKUsArray.pushMap(skuMap)
+    }
+    resultMap.putArray("purchasedSKUs", purchasedSKUsArray)
+
+
+    val relatedSKUsArray: WritableArray = WritableNativeArray()
+    val relatedSKUs = entitlement.relatedSKUs
+    for (sku in relatedSKUs) {
+        val skuMap = skuToSkuDict(sku)
+        relatedSKUsArray.pushMap(skuMap)
+    }
+    resultMap.putArray("relatedSKUs", relatedSKUsArray)
+
+    // For react, provide the most recent active purchase and sku from the arrays
+
+    var lastPurchase: NamiPurchase? = null
+    if (entitlement.activePurchases.count() > 0) {
+        for (purchase in entitlement.activePurchases) {
+            if (lastPurchase == null || lastPurchase.purchaseInitiatedTimestamp < purchase.purchaseInitiatedTimestamp) {
+                lastPurchase = purchase
+            }
+        }
+//            lastPurchase?.let { resultMap.putMap("latestPurchase", purchaseToPurchaseDict(lastPurchase)) }
+    }
+
+    var lastPurchasedSKU: NamiSKU? = lastPurchase?.purchasedSKU
+
+    if (lastPurchasedSKU == null) {
+        val lastPurcahsedSkuID = lastPurchase?.skuId
+        if (lastPurcahsedSkuID != null ) {
+            for (sku in entitlement.purchasedSKUs) {
+                if (sku.skuId == lastPurcahsedSkuID) {
+                    lastPurchasedSKU = sku
+                }
+            }
+        }
+    }
+    if (lastPurchasedSKU == null && entitlement.purchasedSKUs.count() > 0) {
+        lastPurchasedSKU = entitlement.purchasedSKUs.last()
+    }
+//        lastPurchasedSKU?.let { resultMap.putMap("lastPurchasedSKU", skuToSkuDict(lastPurchasedSKU)) }
+
+
+    return resultMap
+}
+
 
 // Convert Java Date to ISO860 UTC date to pass to Javascript
 fun javascriptDateFromKJavaDate(date: Date): String {
